Prevent coil from loading images in full size
This was causing large images to be loaded entirely in memory, which isn't great.
A downside of this commit is that the calculated 'expectedSize' is now incorrect
because it points to a down-sized image by coil. This will now cause SubSamplingImage
to initially load bitmaps for an incorrect size and delay the image's load by a
few frames.